# IGNITE-13293 .NET: Fix enum serialization performance