--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/BenchmarkEnum.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/BenchmarkEnum.cs
@@ -1,0 +1,9 @@
+namespace Apache.Ignite.BenchmarkDotNet.Models
+{
+    public enum BenchmarkEnum
+    {
+        Foo,
+        Bar,
+        Baz
+    }
+}