--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithEnumField.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithEnumField.cs
@@ -1,0 +1,7 @@
+namespace Apache.Ignite.BenchmarkDotNet.Models
+{
+    public class ClassWithEnumField
+    {
+        public BenchmarkEnum BenchmarkEnum { get; set; } 
+    }
+}