--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithIntField.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithIntField.cs
@@ -1,0 +1,7 @@
+namespace Apache.Ignite.BenchmarkDotNet.Models
+{
+    public class ClassWithIntField
+    {
+        public int Foo { get; set; }
+    }
+}