--- a/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryWriter.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryWriter.cs
@@ -904,10 +904,12 @@ namespace Apache.Ignite.Core.Impl.Binary
             _stream.WriteInt(desc.TypeId);
             _stream.WriteInt(val);
 
-            var metaHnd = _marsh.GetBinaryTypeHandler(desc);
-            
             // TODO: This happens on every write - bottleneck. We only need to send a given enum once.
-            SaveMetadata(desc, metaHnd.OnObjectWriteFinished());
+            // Cache this somehow in Marshaller.
+            {
+                // TODO: Fields are always empty for enums, why can't we pass null?
+                SaveMetadata(desc, null);
+            }
         }
 
         /// <summary>