--- a/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryWriter.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryWriter.cs
@@ -904,11 +904,14 @@ namespace Apache.Ignite.Core.Impl.Binary
             _stream.WriteInt(desc.TypeId);
             _stream.WriteInt(val);
 
-            // TODO: This happens on every write - bottleneck. We only need to send a given enum once.
-            // Cache this somehow in Marshaller.
+            var binaryTypeHolder = Marshaller.GetCachedBinaryTypeHolder(desc.TypeId);
+            if (binaryTypeHolder == null || !binaryTypeHolder.IsSaved)
             {
-                // TODO: Fields are always empty for enums, why can't we pass null?
-                SaveMetadata(desc, null);
+                // Save enum fields only once - they can't change locally at runtime.
+                var metaHnd = _marsh.GetBinaryTypeHandler(desc);
+                var binaryFields = metaHnd.OnObjectWriteFinished();
+                
+                SaveMetadata(desc, binaryFields);
             }
         }